ログイン時・未ログイン時のみ操作可能な機能の設定。

+++ routes/web.php
@@ -31,35 +31,35 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, 'index']);
 
 // 求人の新規登録画面を表示する
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // 求人の登録処理をする
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // 求人情報編集画面を表示する
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // 求人情報を編集・更新する
-Route::put('listings/{listing}', [ListingController::class, 'update']);
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // 求人を削除する
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // 求人の詳細を表示する
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 
 // ユーザー登録フォーム画面を表示する
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // ユーザーを新規登録する
 Route::post('/users', [UserController::class, 'store']);
 
 // ログインフォームを表示する
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // ログインする
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 // ログアウトする
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
